Add PayloadPlansConverter trait
