+++ src/main/scala/beam/agentsim/agents/freight/input/GenericPayloadPlansConverter.scala
@@ -1,0 +1,5 @@
+package beam.agentsim.agents.freight.input
+
+import com.typesafe.scalalogging.LazyLogging
+
+class GenericPayloadPlansConverter extends LazyLogging with PayloadPlansConverter {}
